add example data for testing
