+++ arkane/data/CH2CHOOH/r0/IOfiles/r0.ics
@@ -1,0 +1,24 @@
+start_connectivity
+     1-2      1-3      1-6      2-4      2-5     
+     6-7      7-8     
+end_connectivity
+
+start_ics
+    #--------
+    # Stretches (7):
+    #--------
+    1-2         1-3         1-6         2-4         2-5         
+    6-7         7-8         
+    #--------
+    # Bent bond angles (6):
+    #--------
+    1-2-5       1-6-7       2-1-6       3-1-6       4-2-5       
+    6-7-8       
+    #--------
+    # Linear bond angles (2x0):
+    #--------
+    #--------
+    # Improper torsions & Torsions (5):
+    #--------
+    1-4-5-2     1-6-7-8     2-1-6-7     2-3-6-1     4-2-1-6     
+end_ics
