+++ arkane/data/CH2CHOOH/r0/IOfiles/r0.vfit
@@ -1,0 +1,125 @@
+#----------------------------------------------#
+# fitting correlation: (1.0 - r^2) = 5.5e-07   #
+# average abs. errors:                         #
+#    5.8e-01                                   #
+#    2.4e-01 (for points below mean value)     #
+# elapsed time: 1245.3 seconds                 #
+#----------------------------------------------#
+reference -226.3984362 
+const       -     -    +1847.41773
+cos        01     -     -116.48608
+cos        02     -     -361.46338
+cos        03     -     -296.85045
+cos        04     -       +8.00355
+cos        05     -       -0.86581
+cos        06     -      +10.99381
+cos        07     -       -0.33782
+cos        08     -       +0.12122
+cos        09     -       -0.50420
+cos         -    01    +1486.61949
+cos         -    02     +454.80340
+cos         -    03      +66.45607
+cos         -    04       +5.37029
+cos         -    05       +0.77335
+cos         -    06       -0.29414
+cos         -    07       -0.18899
+cos         -    08       -0.13161
+cos         -    09       -0.01420
+coscos     01    01     -148.37682
+coscos     02    01     +282.07038
+coscos     03    01      +19.54587
+coscos     04    01      +15.38158
+coscos     05    01       -7.97061
+coscos     06    01       +2.32854
+coscos     07    01       -1.60420
+coscos     01    02      -48.40197
+coscos     02    02     +168.56720
+coscos     03    02      +87.55364
+coscos     04    02      +50.19849
+coscos     05    02       -1.34533
+coscos     06    02       +1.51403
+coscos     07    02       -1.86457
+coscos     01    03      +10.04162
+coscos     02    03      +28.98124
+coscos     03    03      +40.48145
+coscos     04    03      +27.05970
+coscos     05    03      +11.04601
+coscos     06    03       +2.37467
+coscos     07    03       -0.38704
+coscos     01    04       +5.01304
+coscos     02    04       +9.62542
+coscos     03    04      +19.77049
+coscos     04    04      +13.32235
+coscos     05    04       +7.19094
+coscos     06    04       +2.04611
+coscos     07    04       +0.14653
+coscos     01    05       +1.60137
+coscos     02    05       +2.27457
+coscos     03    05       +5.15439
+coscos     04    05       +4.18065
+coscos     05    05       +3.19128
+coscos     06    05       +1.09007
+coscos     07    05       +0.17333
+coscos     01    06       +0.31904
+coscos     02    06       +0.09099
+coscos     03    06       +0.87021
+coscos     04    06       +0.89440
+coscos     05    06       +0.89433
+coscos     06    06       +0.32646
+coscos     07    06       +0.03900
+coscos     01    07       +0.07452
+coscos     02    07       -0.10076
+coscos     03    07       +0.01315
+coscos     04    07       +0.06229
+coscos     05    07       +0.11783
+coscos     06    07       +0.03372
+coscos     07    07       +0.00408
+sinsin     01    01      +25.59483
+sinsin     02    01     -213.83490
+sinsin     03    01      -42.90067
+sinsin     04    01      -22.79421
+sinsin     05    01       +8.63966
+sinsin     06    01       -2.60099
+sinsin     07    01       +1.38720
+sinsin     01    02      +33.56623
+sinsin     02    02     -223.97636
+sinsin     03    02      -92.67367
+sinsin     04    02      -47.98561
+sinsin     05    02       +1.79597
+sinsin     06    02       -1.51715
+sinsin     07    02       +1.86364
+sinsin     01    03      -18.21220
+sinsin     02    03      -25.71513
+sinsin     03    03      -41.01485
+sinsin     04    03      -26.66434
+sinsin     05    03      -11.30244
+sinsin     06    03       -2.50516
+sinsin     07    03       +0.34994
+sinsin     01    04       -5.70357
+sinsin     02    04       -9.06133
+sinsin     03    04      -18.76348
+sinsin     04    04      -13.36271
+sinsin     05    04       -7.31689
+sinsin     06    04       -2.11505
+sinsin     07    04       -0.16834
+sinsin     01    05       -1.38260
+sinsin     02    05       -1.74686
+sinsin     03    05       -4.93706
+sinsin     04    05       -4.23740
+sinsin     05    05       -3.23474
+sinsin     06    05       -1.12443
+sinsin     07    05       -0.20371
+sinsin     01    06       -0.07396
+sinsin     02    06       +0.00637
+sinsin     03    06       -0.87400
+sinsin     04    06       -0.91726
+sinsin     05    06       -0.91553
+sinsin     06    06       -0.35363
+sinsin     07    06       -0.06918
+sinsin     01    07       +0.00987
+sinsin     02    07       +0.08897
+sinsin     03    07       -0.04981
+sinsin     04    07       -0.08260
+sinsin     05    07       -0.13064
+sinsin     06    07       -0.03667
+sinsin     07    07       +0.00024
