+++ arkane/data/CH2CHOOH/r0/r0.calcs
@@ -1,0 +1,100 @@
+#-----------------------------#
+start_scangeom gaussian
+%nproc=1
+%mem=1GB
+%chk=[Q2DTor_name].chk
+#p hf 3-21g
+   scf=tight NoSymmetry [Q2DTor_MOread]
+   opt=(tight,modredundant)
+
+Geometry optimization at the 2D-PES
+
+0 1
+[Q2DTor_geometry]
+
+ 2  1  6  7 F
+ 1  6  7  8 F
+
+end_scangeom
+#-----------------------------#
+
+
+
+#-----------------------------#
+start_sp0 gaussian
+%nproc=1
+%mem=1GB
+%chk=[Q2DTor_name].chk
+#p hf 3-21g
+   scf=verytight
+   opt=tight
+
+Optimization of a minimum in the 2D-PES
+
+0 1
+[Q2DTor_geometry]
+
+--Link1--
+%nproc=1
+%mem=1GB
+%chk=[Q2DTor_name].chk
+#p hf 3-21g
+   scf=verytight freq=noraman geom=allcheck
+
+end_sp0
+#-----------------------------#
+
+
+
+#-----------------------------#
+start_sp1 gaussian
+%nproc=1
+%mem=1GB
+%chk=[Q2DTor_name].chk
+#p hf 3-21g
+   scf=verytight
+   opt=(tight,ts,calcfc,noeigentest)
+
+Optimization of a saddle point in the 2D-PES
+
+0 1
+[Q2DTor_geometry]
+
+--Link1--
+%nproc=1
+%mem=1GB
+%chk=[Q2DTor_name].chk
+#p hf 3-21g
+   scf=verytight freq=noraman geom=allcheck
+
+end_sp1
+#-----------------------------#
+
+
+
+#-----------------------------#
+start_sp2 gaussian
+%nproc=1
+%mem=1GB
+%chk=[Q2DTor_name].chk
+#p hf 3-21g
+   scf=verytight
+   opt=(tight,saddle=2,calcfc,noeigentest)
+
+Optimization of a maximum in the 2D-PES
+
+0 1
+[Q2DTor_geometry]
+
+--Link1--
+%nproc=1
+%mem=1GB
+%chk=[Q2DTor_name].chk
+#p hf 3-21g
+   scf=verytight freq=noraman geom=allcheck
+
+end_sp2
+#-----------------------------#
+
+
+
