+++ arkane/data/CH2CHOOH/r0/r0.inp
@@ -1,0 +1,74 @@
+#----------------------------------#
+# Torsional information            #
+#----------------------------------#
+start_torsions
+   torsion1         2-1-6-7         # atoms involved in torsion 1
+   torsion2         1-6-7-8         # atoms involved in torsion 2
+   tsigma1          1               # torsional symmetry number of hindered rotor 1
+   tsigma2          1               # torsional symmetry number of hindered rotor 2
+end_torsions
+#----------------------------------#
+# Calculations                     #
+#----------------------------------#
+start_calcs
+   level            hf 3-21g        # the calculation level
+   charge           0               # charge
+   multiplicity     0               # spin multiplicity, shouldn't matter in context of Arkane
+end_calcs
+#----------------------------------#
+# Torsional PES                    #
+#----------------------------------#
+start_pes
+   t1step           10.0            # step in phi1 for scan calculation [degrees]
+   t2step           10.0            # step in phi2 for scan calculation [degrees]
+   symmetry         b               # Symmetry condition for PES: [a,b,c,ab,ac,bc,abc] or none
+end_pes
+#----------------------------------#
+# Fitting details                  #
+#----------------------------------#
+start_fourier
+   weight           0.9             #
+   ignore           0.0             # Set to zero coefficients with smaller absolute value (in cm^-1)
+   # Fourier Terms (Even)           #
+   cos1             1-9             # i values in cos(i*Phi_1)
+   cos2             1-9             # j values in cos(j*Phi_2)
+   cos1cos2         1-7 , 1-7       # i,j values in cos(i*Phi_1) * cos(j*Phi_2)
+   sin1sin2         1-7 , 1-7       # i,j values in sin(i*Phi_1) * sin(j*Phi_2)
+   # Fourier Terms (Odd)            #
+   sin1             1-9            # i values in sin(i*Phi_1)
+   sin2             1-9            # j values in sin(j*Phi_2)
+   cos1sin2         1-7 , 1-7            # i,j values in cos(i*Phi_1) * sin(j*Phi_2)
+   sin1cos2         1-7 , 1-7            # i,j values in sin(i*Phi_1) * cos(j*Phi_2)
+end_fourier
+#----------------------------------#
+# Search and Opt stationary points #
+#----------------------------------#
+start_statpoint
+   tolerance        1.0             # step (in degrees) to explore torsional PES when looking for CPs
+   freqscal         1.000           # scaling factor for frequencies
+end_statpoint
+#----------------------------------#
+# 2D-NS Hamiltonian                #
+#----------------------------------#
+start_tor2dns
+   dijvar           yes             # yes (dij not constant) or no (dij constant)
+   kmax             100             # check 2013-JChemPhys_138_134112, eq (14)
+   maxeigen         1e4             # threshold for H eigenvalues (in cm^-1)
+end_tor2dns
+#----------------------------------#
+# Partition functions              #
+#----------------------------------#
+start_rovibpf
+   interpolation    fourier         # fourier or spline order (1,3,5)
+   integrationstep  1.0             # integration dphi
+end_rovibpf
+#----------------------------------#
+# Working temperatures             #
+#----------------------------------#
+start_temperatures                 #
+   100.0   150.0   200.0          #
+   250.0   300.0   400.0          #
+   500.0   700.0  1000.0          #
+   1500.0  2000.0  2500.0         #
+end_temperatures                   #
+#----------------------------------#
