+++ arkane/data/CH2CHOOH/r0/r0.xyz
@@ -1,0 +1,10 @@
+8
+reference geometry for S02
+C    0.755353   0.552739   0.025245
+C    1.491563   -0.529139   -0.064237
+H    1.176388   1.534477   0.085291
+H    2.556556   -0.431497   -0.078136
+H    1.117489   -1.530777   -0.126579
+O    -0.62445   0.692591   0.061762
+O    -1.398931   -0.544415   -0.009169
+H    -0.746876   -1.261019   -0.071635
